Specify globalCIDR while joining cluster
Enhance 'subctl join' to be able specify globalCIDR
for each cluster while joining. The logic implemented is

if globalcidr specified by user
  if globalcidr is preconfigured
    take pre-configured value
  else
    error if it overlaps with other clusters
    take specified value

if not specified by user
  if globalcidr is preconfigured
    take pre-configured value
  else automatically assign value

It also checks
* if the specified CIDR is valid
and that both globalcidr and globalnet cluster
size are not specified.
* if globalnet-cidr or globalnet-cluster-size
is specified even if globalnet is not enabled
and ignores its value.

Closes: #298
Signed-off-by: Janki Chhatbar